Delete util files.
