+++ src/av/video.h
@@ -51,4 +51,18 @@ void utox_video_thread(void *args);
 
 void postmessage_video(uint8_t msg, uint32_t param1, uint32_t param2, void *data);
 
+
+// Color format conversion functions
+
+void yuv420tobgr(uint16_t width, uint16_t height, const uint8_t *y, const uint8_t *u, const uint8_t *v,
+                 unsigned int ystride, unsigned int ustride, unsigned int vstride, uint8_t *out);
+void yuv422to420(uint8_t *plane_y, uint8_t *plane_u, uint8_t *plane_v, uint8_t *input, uint16_t width, uint16_t height);
+void bgrtoyuv420(uint8_t *plane_y, uint8_t *plane_u, uint8_t *plane_v, uint8_t *rgb, uint16_t width, uint16_t height);
+void bgrxtoyuv420(uint8_t *plane_y, uint8_t *plane_u, uint8_t *plane_v, uint8_t *rgb, uint16_t width, uint16_t height);
+
+// TODO: Documentation.
+void scale_rgbx_image(uint8_t *old_rgbx, uint16_t old_width, uint16_t old_height, uint8_t *new_rgbx, uint16_t new_width,
+                      uint16_t new_height);
+
+
 #endif
