+++ src/cocoa/interaction.m
@@ -8,9 +8,9 @@
 #include "../friend.h"
 #include "../main.h"
 #include "../messages.h"
+#include "../text.h"
 #include "../tox.h"
 #include "../ui.h"
-#include "../util.h"
 #include "../utox.h"
 
 #include "../av/utox_av.h"
