+++ src/command_funcs.c
@@ -4,7 +4,6 @@
 #include "groups.h"
 #include "logging_native.h"
 #include "tox.h"
-#include "util.h"
 
 // FIXME: Required for UNUSED()
 #include "main.h"
