+++ src/commands.c
@@ -4,7 +4,6 @@
 #include "flist.h"
 #include "logging_native.h"
 #include "main.h"
-#include "util.h"
 #include "ui/edits.h"
 
 struct Command commands[MAX_NUM_CMDS] = {
