+++ src/filesys.h
@@ -43,4 +43,10 @@ bool utox_remove_file(const uint8_t *full_name, size_t length);
 
 bool utox_move_file(const uint8_t *current_name, const uint8_t *new_name);
 
+/* read a whole file from a path,
+ *  on success: returns pointer to data (must be free()'d later), writes size of data to *size if size is not NULL
+ *  on failure: returns NULL
+ */
+void *file_raw(char *path, uint32_t *size);
+
 #endif
