+++ src/groups.c
@@ -3,7 +3,7 @@
 #include "flist.h"
 #include "logging_native.h"
 #include "main.h"
-#include "util.h"
+#include "text.h"
 
 #include "ui/edit.h"
 #include "ui/scrollable.h"
