+++ src/windows/os_video.c
@@ -4,7 +4,6 @@
 
 #include "../av/video.h"
 #include "../logging_native.h"
-#include "../util.h"
 
 #include "../../langs/i18n_decls.h"
