+++ src/xlib/gtk.c
@@ -4,8 +4,8 @@
 #include "../flist.h"
 #include "../friend.h"
 #include "../logging_native.h"
+#include "../text.h"
 #include "../tox.h"
-#include "../util.h"
 #include "../utox.h"
 
 #include <dlfcn.h>
