+++ src/xlib/main.c
@@ -8,7 +8,6 @@
 #include "../logging_native.h"
 #include "../theme.h"
 #include "../tox.h"
-#include "../util.h"
 #include "../utox.h"
 #include "../av/utox_av.h"
 #include "../ui/draw.h"
