+++ src/xlib/video.c
@@ -1,6 +1,5 @@
 #include "main.h"
 #include "../logging_native.h"
-#include "../util.h"
 #include "../av/video.h"
 
 #include <stdio.h>
